Skeleton for 3 python packages: walt-client, walt-server and walt-node.
walt-server is a daemon that should be running on the server.
walt-client is the command line tool that allows to control and monitor
the testbed. It should also be installed on the server.
walt-node should be installed on the nodes (RPis).

There is a sub-directory for each of these 3 python packages.

A setup.py script should be provided for each of these packages, in
order to easily build the python package and upload it on PyPI.
This commit introduces a 1st version of client/setup.py and
server/setup.py.

Skeleton command line interface for the client and the server
are implemented.

Finally, a platform discovery function is implemented on this
skeleton. The client can issue the following requests:
$ walt platform print	# show the table of platform devices
and
$ walt platform rescan  # update this table by rescaning the network