Skeleton extended with a 1st client -> server -> node command: blink.
When the client types
$ walt node blink <node_ip> 60
the request is transmitted to the server and then to the node at
<node_ip>.

This can be tested by running walt-node-daemon on the server with the
--fake command line switch, to simulate that it is running on
a real walt node, and use <node_ip> = <server_ip> in the command above.

Apart from the management of this 'blink' command, this commit
involves a skeleton for walt.node.daemon, and code refactoring
using a new subpackage called walt.common.