Associate docker images to WalT nodes and export them through nfs.
The database contains a new table called 'nodes', that allows
to associate the name of an image (i.e. an operating system image)
to a given node.

An image is actuelly a docker image. It should be available either
locally or on the docker hub.

Manipulating the new 'nodes' table is not yet implemented in the
walt client.
However, the mounting / unmounting of an image is implemented.
When mounted, the image is also exported through NFS.
The server reads the table 'nodes' on startup, and automatically
mounts the images needed. When it stops, it un-mount them. Here is a
sample run (with an entry manually added in the table 'nodes'):
~# python
[...]
>>> from walt.server.daemon import run
>>> run()
Mounting waltplatform/rpi-debian... done
Initializing... Done.
^CInterrupted.
Un-mounting waltplatform/rpi-debian... done
~#

The mountpoint of an image is at:
/var/lib/walt/images/<image_name>/fs
e.g. in the previous example:
/var/lib/walt/images/waltplatform/rpi-debian/fs

This commit required to ensure that images are unmounted whenever
the server daemon stops, in order to keep the system in a clean
state.
This required some refactoring of the existing code.

Next steps:
- implement 'nodes' table management and make it available
  for the walt client
- update the dhcpd configuration in order to let the nodes
  boot the system embedded in the image.