Manage default (walt-node-)image.
The dhcp server as a pool of addresses for unregistered nodes.
The default (walt-node-)image is the image that will be affected
to them.

This commit introduce a new table called 'config' in the server db,
which should be used to manage user-modifiable configuration items.

A 1st item called 'default_image' is introduced.
The first time the server is started, this item is not found. In
this case the server affects one of the available image names.
At each startup, the server now mounts all images attached to at
least 1 node, plus this default image.

Next step:
- add command line options to let the user modify it.