Auto-register new nodes.
When a walt node starts, walt-node-daemon connects to the server
and calls the remote method register_node(<node_ip>). If the
IP is not known in the database yet, the topology is rescanned
which causes the new node to be registered and the current default
image to be associated to it.