Add images & nodes management commands.
This adds the following user commands:
- walt image list
- walt image set-default <image-name>
- walt node list
- walt node set-image <node-name> <image-name>

An image is an operating system deployable on a WalT node.
Internally it is a docker image. The kernel must be stored at path
/kernel in this image (it may be a symlink to another file of the
image).
The DHCP server will direct new nodes (nodes whose mac address
is seen for the 1st time) to a default image. This default image
can be configured using the 2nd command listed above.
The 3 other commands should be self-explanatory.