SIGTERM: Handle appropriately.
Background: the command "service walt-server stop" causes
a SIGTERM signal to be sent to walt-server-daemon.
(and then a SIGKILL signal if it is still alive after
a little time.)
When SIGTERM is received, we now raise a SystemExit exception.
This allows to appropriately catch this event within 'with'
or 'try' statements, and thus leave the process in a clean
state.