Only consider docker repositories called 'walt-node' for node images
The list of available walt node images is now established by looking
for repositories called 'walt-node' both locally and on the docker
hub.
Such an image could be for example:
- waltplatform/walt-node:board_rpi-os_debian (available from the
docker hub), or
- local/walt-node:board_rpi-os_redhat-osversion_5.0 on the local
disk.

The limitations of the docker tagging feature helped in choosing
this solution.

We may improve the user interface regarding this in the future
(since all displayed image names contain 'walt-node', we may
remove that; for the tag, if we stick to a naming scheme such as
the one taken in the examples above, we could parse image
'attributes' there and display them in a user-friendly way).