Introduce an EventLoop class.
Standard RPyC servers only listen on the RPyC socket. Whereas the
WalT daemon will have to listen both on the RPyC socket and on TCP
connections to manage logs.
This commit introduces the EventLoop class, and adapts the
SimpleRPyCServer class to use this class.
It will allow to listen to several file descriptors simultaneously,
by registering additional listeners on this event loop.