The event loop is now created outside of the daemon.
This change makes the server code cleaner because
the daemon is mostly implemented in the walt.common
package, and thus retrieving an object created there
to manipulate it in the server code is not very clean.

The node daemon code was also adapted since it uses
the same base class.