Experience logs: Preliminary implementation (server-side)
The server daemon now listen to one more TCP port
that will receive walt experience logs.
Before connecting and sending logs, one must first
register a new 'log stream' and give its name and
sender mac address; the result of this procedure
is a new allocated 'stream id'.
Then, one can open a connection to the TCP server,
and send the stream id as the first serialized object.
This will allow the server to link this connection to
the given stream id. The next serialized objects sent
on the TCP connection are log records.
Objects sent through the TCP connection must be
serialized using pickle.
When log records are received at the server, they go
through a 'hub'. The hub is linked to a set of 'handlers'.
Each log record will be sent to all handlers of the hub.
A first handler has been implemented. It is responsible
for storing the log records in the sqlite database.
It is possible to dynamically add handlers to the hub.
This will be used when the command-line client wants to
display the logs in real time. (this is not implemented
yet.)