Experience logs: Preliminary implementation (node-side)
This commit provides a new command called walt-monitor, that
should be run on walt nodes in order to monitor a given command.
For example:
$ walt-monitor ./bandwidth-test eth0
This will run './bandwidth-test eth0' in a subprocess and monitor
this subprocess. Currently stdout and stderr streams are monitored,
but we may add more in the future (CPU/mem ressource consumption,
lifecycle information such as start/end times and return status,
etc.)
The collected information (i.e. stdout / stderr for now) is sent
to the walt-server as logs. (Thus it can be displayed using
'walt logs show' as described in the previous commit.)