Let the client run interactive prompts.
This commit allows to run interactive command shells at the server
and let the client interact with them (i.e. the stdin of the client
is more or less piped to the stdin of the shell, and vice-versa for
stdout).
Let the server deal appropriately with this was actually more complex,
because the server runs in one loop only, so the management of the
prompt had to be plugged on the event loop.
The handling of RPyC was refactored to handle one request at a time
in the event loop (instead of one client at a time, which meant that
we could not return to the loop before the client was disconnected).

The user command 'walt advanced sql' is a first application of such
a prompt:

$ walt advanced sql
psql (9.4.1)
Type "help" for help.

walt=> select * from devices;
        mac        |       ip       |    name     | reachable |  type
-------------------+----------------+-------------+-----------+--------
 6c:b0:ce:18:d5:5d | 192.168.152.2  | main-switch |         1 | switch
 c0:3f:d5:67:10:a6 | 192.168.152.1  | walt-server |         1 | server
 b8:27:eb:24:b4:a1 | 192.168.152.14 | rpi-ED      |         1 | rpi
(3 rows)

walt=> \q

Remote shell ended.
$