New docker-run based prompt (for image modification)
A new command is provided:
walt image modify <image-name>
This starts a remote shell based on a 'docker run ...'
command running on the server.
This command is not fully implemented, only the prompt
is handled. The modifications made from the prompt
should be saved to the image or to a new image. This is
not implemented yet.

The management of remote prompts (running on the server
and displayed on the client) was much improved.
The main trick is that we must change the tty settings
of the client in order to avoid interpreting the escape
sequences (ctrl-r etc.), and instead pass them to the
server unchanged.

The TCP server that was used for faster processing of
logs (comparing to using RPyC for that) has been refactored
to make it more flexible. The management of remote prompts
is now also plugged on this TCP server.
(Whether we should also manage RPyC connections as a module
of this TCP server mechanism is an open question. For now
we have 2 separate TCP ports.)