New command 'walt node deploy' (instead of set-image)
'walt node deploy' allows to deploy an operating
system image on a walt node.
Here is the process:
- poweroff the node
- associate the node to the image
- poweron the node

The command 'walt node set-image' has been removed.
We consider that if the user wants to associate an
image to a node, then he will want to reboot it
for this request to take effect.