Mount-time procedure to adapt the node image to the WalT setup.
The walt-node package now installs a new command 'walt-node-install'.
This command is called when the image is mounted, and thus allows
to adapt the image to the WalT environment when the image is
mounted (e.g. refer to the server ip in the file /etc/ntp.conf of
the node).