walt node shell: remote shell session to a walt node
This reuses the work on interactive prompts (see relative
commits).
We provide root access to nodes with no password:
security is useless in this case because nodes are very
temporary environments, and any change will be forgotten
when they are rebooted. (Moreover the WalT network is
separated from the rest by a dedicated VLAN, with no
routing at the server.)
In order to let ssh connect with no password, the pubkey
of the server is copied to [image]/root/.ssh/authorized_keys
when the image is mounted.
Furthermore, the ssh key used for host-authentication is
overwritten by a constant one. Otherwise the server would
complain about the remote host key change, when a new image
with different keys is deployed on the node.