Fix dhcpd.conf generation.
A query to the database gives us the list of devices.
In this loop, if a given device is a node, a query
to table 'nodes' is also performed.
We have to fetch all the results from the 1st query
before performing the additional ones.
This was not the case, thus the results of the 1st
query were truncated after the 1st node was processed.