Fix log channels handling at server side.
When opening a connection to the server to send logs,
one has to send the log stream label chosen.
the server was reading this label and then was waiting
for the first log record, without returning control to
the event loop.
If no log record is ever sent to this logstream (e.g.
a log stream for stderr), the server was blocked.