Re-worked walt-monitor and related log handling.
walt-monitor was a python script. On a raspberry pi,
python scripts are very slow to be started.
'walt-monitor echo test' was taking more than 2 seconds.

We now have most of the functionality implemented
by walt-node-daemon, and walt-monitor is now a shell
script.
The communication between these two is handled by
FIFOs.

walt-node-daemon waits for requests on
/var/lib/walt/logs.fifo
The first request implemented is:
MONITOR <stream_name> <fifo_file>
Upon receipt of a request of this kind, walt-node-daemon
will open a new log connection to the server, with
name <stream_name>, then it will start reading <fifo_file>
and forward any log line through this log connection.

Thus, walt-monitor just has to:
- creates 2 fifo files, to handle stdout and stderr of the
command that must be monitored
- send the 2 relevant MONITOR requests to /var/lib/walt/logs.fifo
- start the specified command with stdout and stderr redirected
to the relevant FIFOs.