Add restrictions on device name.
Since the device name is used as the node's hostname
in dhcpd.conf, the device name must be valid in this
context.
For example some protocols such as avahi (zeroconf)
will not work if it contains an underscore.
This commit verifies the new device name when using
'walt device rename', and changes the default name
when a new device is discovered.