ev_loop: handle planned events. db: periodic auto-commit.
The event loop is now able to fire events at a given time,
and optionally repeat them in a periodic fashion.
This is used for a periodic commit of the database. Logs
may generate a high rate of events, therefore we do not
commit to the database after each event, and use this
periodic commit instead. This allows to reduce the server
load.