Improved client config.
We add comments in the json file ~/walt.conf in order to be able to
describe each item in more details. Since it is not a standard
feature in json, we manage these comments manually.
We also ensure that the items will be printed in a logical order
in the json file.
We provide an attribute to specify is a given missing attribute
should be prompted to the user (or if the default value should
be chosen silently).