First Implementation of search and clone.
This also involves support for asynchronous blocking
(i.e. lenghty) operations in the server.
An object of class walt.server.blocking.BlockingTasksManager
is associated with a separate thread that performs these
lengthy operations. The main thread is woken up when the result
is ready because the result is sent through a pipe and thus
it is handled appropriately through the event loop mechanism.
This mechanism is used for the new walt image search and
walt image clone operations.

Finally, a skeleton for the other walt image xxx operations
is in place.