Implemented walt image show.
The object responsible for image management operations
has been renamed from NodeImageStore to NodeImageManager.

A new object called NodeImageStore (again) has been
introduced. It is a container of the local docker images
related to WalT (images whose repo is 'walt-node').
Working with this object avoids calls to the docker API,
thus it acts as a kind of cache. When an operation changes
the set of docker images contained (such as
walt image [clone|rename|copy|remove]), one must call its
method refresh().

The clone operation has been modified to call this refresh()
method of the NodeImageStore.

The show operation has been implemented: it reads the content
of the NodeImageStore and looks for images whose username is
the same as the requester's.