walt image clone: allow overwriting an existing image.
If an image with the same tag exists in the working set,
a message is printed to tell the user that it should use
option --force.
If this option is specified, the process continues and
the existing image is overwritten. If it was mounted,
it is unmounted and re-mounted with the new version.
As a side effect, any node using this image is automatically
rebooted.