Improved new node / default image management
Old behavior: we had a default image, configurable
by the client, that was used for nodes that we met
from the first time.
The appropriate conf option was specified in
the group of unregistered devices in the dhcp conf
file.
The first time the server was booted, the default
image was set to the 1st walt docker image found
on the server.
A node was registered (with attributes such as its
ip address stored in the db) either when it was detected
by the topology exploration process, or by a request
sent from walt-node-daemon (thus once the node is
fully booted).
Problem: having a single default image whatever is the
walt node, we could not handle several kinds of walt
node hardware.

New behavior:
* dhcpd sends no image configuration for unknown nodes
* new nodes detect in their 1st boot phase that this
  option is missing. Thus they know they are a new node,
  and they send a registration request to the TCP server
  module of the walt server. This request includes the ip,
  mac and node type (e.g. rpi) information.
* the server looks for the default image which is, for a
  given node type, waltplatform/walt-node:<type>-default,
  and pulls it from the docker hub if needed. Then it
  registers this node in the database together with the
  association to this default image, and updates dhcpd.
* when the new node makes a new attempt, it gets the option
  about the walt image to boot. Thus the usual boot up
  procedure can continue.

Notes:
* the client option and related server code to specify a default
  image has been removed (not needed anymore).
* the node registration process that was issued by walt-node-daemon
  (once the node is fully booted) has been removed (not needed
  anymore).
* passing requests to the TCP server module has been simplified
  in order to allow a node in its 1st bootup phase to send a
  request easily (e.g. using bash and nc, whereas before the
  request was python-pickle encoded).
* the git repo walt-rpi-debian has also been updated in order to tag
  and publish rpi-jessie as rpi-default.