Fix rare synchro/buffering issue with interactive shells.
Using walt.common.io.unbuffered() (allows to re-open a
python file with buffering disabled) may cause data loss
if current data was ready to be read at the time one
performs this call.
We now send a message 'READY\n' from server to client
to avoid that:
* the client connects to the TCP server
* it sends a SHELL request id
* this request id is catched by the event loop, which
  create the appropriate listener to handle this shell
* the listener reopens the socket connection to the
  client by using unbuffered()
* the listener sends the 'READY\n' message to the client
* the client was blocked at this point, waiting for
  this message, so it is unblocked
* the client sends the rest of the needed data (what's
  the <ip> to ping, etc.)

Before, the client was sending the rest of the data
immediately after the request id, which was causing
the issue described above.