Improved node & device detection
Before, 'walt device rescan' caused all topology information
(i.e. where (switch mac & port) a given device is connected)
to be discarded, and then updated with new data.
This caused stability issues, such as not being able to reboot
the poe port of an unreachable node (because the information
about the poe port to reboot had been lost).
Now the information is not removed, it's just updated.
If a given network device was not found by the rescan, its
'reachable' flag is set to 0, but the information about where it
is connected remains. The only way this information could be
lost is to replace this device with another equipment
connected on the same switch port. In this case, 'walt device show'
will report the previous device as 'replaced' and suggest to use
'walt device forget'.
Commands 'walt device [show|tree]' and 'walt node show' were
improved to reflect this.
Finally, nodes notify the server at each bootup. For now, this
causes their 'reachable' flag to be set to 1 (if ever it was 0).
This notification could be used for other purposes in the future,
such as logging this bootup event.