TCPServer fix related to file buffering.
The TCP server accepts new connections, creates a file object
from the connection socket for easier I/O, read the request ID
and creates the appropriate listener, and registers this
listener in the event loop.
The file object mas created by calling socket.makefile(), thus
this file object was buffering input data. Thus if the client
sends other data shortly after the request ID, this data is
read at the same time and stored in the file object buffer.
Since the event loop works by polling the initial file descriptor
(the socket in this case), it will not detect any input data.

This issue was solved by specifying a buffer size of 0 when
calling socket.makefile().