walt log show: added option --nodes <set-of-nodes>
The user can now select which logs to display according
to the node that generated them.
Valid options are the keywords 'my-nodes', 'all-nodes',
or a coma-separated list of nodes.
If unspecified, the default value is 'my-nodes', which
means that only the logs originated by a node deployed
with an image of the user will be displayed.