Node's logs.fifo: Handle the 'LOG' directive.
One may now register a log-line with a given logstream
name by sending
LOG <stream-name> <log-line>
to file /var/lib/walt/logs.fifo.

Internally, the node daemon establishes a stream to the
server for each logstream name. It maintains a cache
of these. If no more logs are recorded for a given streamname
after a few seconds, the stream is closed (it could be
reopened later if needed).