Added a user-friendly curses-based server console.
This console will be displayed when the server boots and will
help the user setup up the testbed on first boot.
A command walt-server-console allows to start it.
The communication from walt-server-daemon to this console
is done through a fifo at /var/lib/walt/ui.fifo.
The console is implemented at walt.server.ui.console,
and the server is sending requests to the console by using
an instance of class walt.server.ui.manager.UIManager.