Improved handling of node registration.
Until now, the image store was just mirroring the local docker images.
Now, we have a new table in database called 'images' and the image
store relies on it.
In this table, we have the docker fullname of the image and a boolean
property called 'ready'. ready is false when an image is being
downloaded (with a docker pull), true otherwise.
When a node registration request comes in, the server checks if the
default image for this node type is already available. If not, the image
is registered anyway, but with ready=false. Then the image is pulled
and ready is passed to true.
As a consequence, 'walt image show' is now able to properly explain
to the user that a given node is 'detected but not ready yet' because
'its image is being downloaded'.
'walt device show' also had its messages clarified: nodes which do
not have any topology information are either new devices or devices that
have been replaced.