Improved handling and user feedback about network init.
The process now checks whether we manage to get a DHCP offer
for the server or not. We do not go further in the init
procedure until we get an IP address.
The user is warned about what is happening during this time,
and textual explanation about what is needed is printed
on the server console.
The network diagram displayed in the console was clarified
about
- port 1: the externally managed VLAN (where we send DHCP requests)
- port 3 and more: the managed walt network.