Server console: update in case of long topology scan
When scanning the topology on server startup, there may
be devices that did not get an IP on the walt network yet,
because the DHCP server running at the server was just
started.
When waiting for these devices to request an IP, and then
waiting for the LLDP protocol to propagate this change,
we sometimes wait for a long time.
Thus, when such a device is detected, we call
ui.task_running() every 1s in order to show on the server
console that we are still alive.