More user-friendly handling of PoE connectivity issues.
In order to reboot or deploy a node, we need to know on which
PoE switch port it is connected. This corresponds to a record
in db table 'topology'. There may be some delay before LLDP
tables are updated and this information is propagated.

In case the user requests such an operation, if this information
is not available we automatically perform a new topology scan and
recheck. If the information is still not available, we indicate
the issue to the user and suggest to retry after a few minutes.