Proper handling of 'free' nodes.
'Free' nodes are nodes that have booted the default image.
Internally, the image owner in this case is 'waltplatform'.

Before, there was no special handling of this, thus these
nodes were considered to be used by another user called
'waltplatform'. They were displayed as such in 'walt node show',
and causing warnings when trying to reboot or deploy an image
on them (i.e. are you sure, it seems to be used by someone else?).

Now they are displayed in a separate category in 'walt node show',
and no longer generating warnings when trying to use them.