Added 'walt node wait': wait for nodes to be ready
This command allows to wait for a set of nodes to be booted.
If their flag 'reachable' is 1 in db, we consider they are
ready. Otherwise, the server records this client request
and will unblock it when the bootup events of all involved
nodes have been received.

The client blocking is implemented by a read() on a ResponseQueue
object. The server writes something on it when appropriate.

The server logic is mostly implemented in module
walt.server.node.wait.