Mount images using docker's internal aufs layers.
In previous versions of WalT, we were running a docker container
in order to expose its filesystem root using NFS.
This was possible with docker<=1.3 because the root filesystem
of a docker container is visible in the host OS.
This is not the case anymore with more recent docker versions.
That's why we were blocked to docker 1.3.
Recently, docker<=1.5 versions have been obsoleted and access
to the docker hub is being discarded for these older versions,
which forced us to find a solution quickly.

With this commit, we do not start a container any more when we
want to mount an image. Instead, we read the configuration of
the image, retrieve the various filesystem layers it is made of,
and manually mount these layers, independently of the docker
engine.