Fix VLAN isolation (WalT vs external VLANs).
The network link between the main switch and the server now
holds tagged packets only, with either VLAN 1 (WalT) or VLAN
169 (external).
(Before, only VLAN 169 was tagged, VLAN 1 was not.)
However, in the setup procedure we send untagged packets in order
to be able to communicate with an unconfigured switch.
This commit also introduces a technique that causes a given switch
to replay its DHCP bootup procedure. It is based on an SNMP query,
see walt.server.snmp.dhcp. When setting up a new testbed, the main
switch may get an IP address from the external DHCP, and the lease
time may be long. The switch will not renew this address (and get
an appropriate address in the WalT subnet) before this lease time
is over. We detect such an issue when we are discovering the topology
and we see that one of the neighbors is a switch and has an IP
outside of the WalT subnet. In this case we use the technique
mentionned above.