Fix walt image cp sometimes not working.
A docker container is started with a tar command allowing
to receive files from stdin.
The code waits for this container to be started, then waits
for its completion (i.e. docker wait), then commits the
container changes into the docker image.

Actually we were just waiting for the container to exist,
thus it was sometimes not running yet. So 'docker wait'
was returning immediately (i.e. the container was already
'stopped' because it was not started yet), and the commit
was occurring before the changes were really done in the
container.
We now really wait for the container to be started.

For safety, sync commands are also added after the tar
is done.
And the popen object running 'docker run' was a local
variable, it is not the case anymore: we save it as an
attribute of the eventloop listener in order to ensure
that it will not be garbage-collected (just in case it
would close the stdin/stdout streams in this case...)