Stop the DHCP client on cleanup.
Since the server code starts the DHCP client, it should
also stop it.
Moreover, for an unknown reason, if we stop the server process
and leave the DHCP client alive, the client starts listening
on TCP port 12347, which is the port the WalT server uses for
its RPyC server. Therefore restarting the daemon fails unless
we kill the DHCP client.
By killing the client on cleanup, we fix this.
Note that even if we stop the DHCP client, we prevent it from
removing the network configuration associated to the lease
(IP address and default route), otherwise it would prevent us
to reach the server through the external network when the WalT
daemon is stopped.