Fix walt node shell file removal not working.
WalT images are made of several layers of the aufs (union)
filesystem. walt node shell creates one more layer in an image.

1 layer is made of a directory where added files are created,
and deleted files are registered by using a special machanism
called "whiteouts". If a file is marked as whiteout, it should
not appear on the union mount.

However, we were mounting images by specifying branches with the
"<branch_path>=ro" syntax in the mount command. The man page
of mount.aufs says that in this case aufs does not look for any
whiteout in this branch. Thus the "deleted" files were still visible
in the union mount.

This commit replaces '=ro' by '=ro+wh', which stands for 'read-only
branch with whiteouts'.