Handle optional image features.
Images may implement optional features. These are
declared in /etc/walt/image.spec in the image filesystem,
and each feature in linked to a script allowing to enable
this feature when the image is mounted.

The server should also provide a file /etc/walt/server.spec
that defines the features implemented on its side.

Features implemented both by the server and the image are
enabled when the image is mounted.
For example, ptp may be enabled if the image provides a ptp
client configuration and the server runs a ptp server daemon.
Thus the image may declare this optional feature and provide
a script /bin/enable-ptp.sh that will disable NTP and activate
its PTP client.
A feature that is declared on one end only will not be enabled.
If a spec file is missing, we consider no optional feature is
implemented.
This should allow an old server installation to run new images,
a new server to run old images, etc.

The walt-node-install code is responsible for checking this
and running the scripts allowing to enable the matching features.