Rewrote 'walt image clone' server procedure.
In the past, several bugs have been found in the implementation of
walt image clone.
This commit provides a new implementation of it. Based on several
variables (whether an image with this name already exists on the
server, or in the user's working set, whether the name of the user
in the clonable link is the same as the current user, etc.) we
select an appropriate workflow (a list of functions) and execute
it.
Check out the comments at the top of server/walt/server/images/clone.py
for more info.