Fix walt node run when not on a tty.
The WalT client was trying to get terminal width from stdout.
This was failing when stdout in redirected.
The client now notifies the server whether stdin and stdout
are a tty. Depending on the command (e.g. walt node run/shell vs
walt node cp) a tty environment is preferred or not. Given
these two parameters, the server starts the backgroup process
using either forkpty() or popen().