Manage docker authentication.
This commit exposes a new function to the client allowing a
user to login to the docker hub.
This will allow to implement 'walt image publish' (hopefully
next commit).

In order to avoid transmitting the password in clear text,
a diffie-hellman process allows to generate a symmetric key that
is used to encrypt and decrypt it using the blowfish algorithm.
(These 2 algorithms were added in walt.common.crypto by a
previous commit.)

On the client side, the password is saved in walt.conf. It is
encoded there using a simple xor-based encoding (just to avoid that
someone could read it on the screen). The security is actually
based on the access rights of walt.conf, which is only readable
by its owner. This design choice has been made because docker
stores authentication data in .docker/conf.json in the same way.

The configuration file management has been completely rewritten,
in order to:
1) allow to read/write encoded passwords
2) make sure that the user specifies valid conf parameters

Until now, point 2 was not so mandatory. We were allowing users
to use WalT with no docker hub account (since a docker hub account
is only needed when pushing images), thus any username was accepted.
And a mistake in the 'server' parameter would also have little impact,
except for causing connection failures before this parameter is
correctly set.

Now, we want to push images, so we have to ensure that the WalT user
specifies the username and password of his docker hub account.
As a result, each missing configuration parameter is now asked for
and validated using an appropriate validity check:
* 'server' parameter -> check connection to the walt server daemon
* 'username' and/or 'password' -> check login to the docker hub
For performance reasons, the parameters already set in walt.conf
are not rechecked each time the walt tool is started.