API introspection mechanism.
This introduces a mechanism to explore the API of server,
client and node components.
This API must be declared with @api and @api_expose decorators.
A subset of this API was decorated, for testing purpose.

Usage:
$ dev/api/explorer.py walt.server.daemon
[API prototypes are displayed.]
$ dev/api/explorer.py walt.client.client
[API prototypes are displayed.]
$ dev/api/explorer.py walt.node.daemon
[API prototypes are displayed.]
$

The trickiest part of this mechanism consists in altering the
import system in order to inspect only the python source files,
not any installed python package. Otherwise, in order to explore
the node API for example, we would need to install the walt-node
package and its dependencies on the development machine, which
would be cumbersome.