Automatic versioning of packages
Using dev/version/updater.py, we update:
* the file <subdir>/info.py
* the file common/walt/common/versions.py

<subdir>/info.py contains a directory of package metadata
that is imported in setup.py.
Since it is just a directory, this file is easy to generate
compared to a standalone setup.py file.

We automatically upgrade package versions given the upload
number and the API changes.

We have 2 new sub-packages: walt-client-selector and
walt-node-selector. They allow to select an appropriate
version of walt-client (resp. walt-node) in order to be
compatible with the server:

$ pip install --upgrade walt-client-selector==<server_api>.<client_api>.*
will automatically select and install a version of the client
able to satisfy the API constraints specified.

$ pip install --upgrade walt-node-selector==<server_api>.<node_api>.*
will do the same thing for walt-node.