Work with client-server and node-server APIs
Since our preliminary work on versioning, we had intruduced the concept
of API versions, and we were looking at three different APIs:
* server API (called by the client and the nodes)
* client API (called by the server)
* node API   (called by the server)

We now work on 2 API sets instead:
* CS API (client and server related APIs)
* NS API (node and server related APIs)

Handling the compatibility between the client and the server
is not a big deal. Between the server and the nodes, it is much more
complicated. With the previous logic, changing the way the server and
the client communicate would many times alter the server API. Considering
the server API is altered means we also have a compatibility break on the
other side (server - nodes).
With the new logic, such a compatibility break is not propagated on the
other side.

In the code, we separated the server API in two separate sections
walt.server.daemon.cs and walt.server.daemon.ns.

walt.server.daemon was a module before, it is now a package and its code
is spread in the files of the server/walt/server/daemon/ directory.
The run() method is now at walt.server.daemon.daemon.run().