Fix walt node deploy my-nodes default
Since we are releasing the nodes we own, the set
of nodes we own is altered by the command.
Because of this we were powering off the nodes
and not powering them on back.
The set of nodes to work on is now saved at the startup
of the command.