Implemented blocking server thread as an EvTread
The daemon, which launches the 2 server threads (main and blocking)
also connects them.
A ThreadConnector class has been implemented in walt.common.thread
for this purpose.
Important note: these connectors are involved in the event loop.
It is mandatory to connect() the connectors before registering the
related listeners in the event loop, otherwise no file descriptor
will be initialized.
Thus the daemon's workflow is (see walt.server.daemon.run):
1- Create the threads -- each thread shoud instanciate its connectors
   in its __init__() method
2- Connect the threads by linking their connectors
3- Start the thread manager -- this will start all threads. Each
   of them will start by calling its prepare() method. The registration
   of objects listening on a connector in the event loop should be
   done there.