Introduce new server thread: ServerHubThread
This new thread should be responsible for all interactive operations,
Such as RPyC client or node requests, interactive shell sessions,
etc.
When the hub thread receives an operation (e.g. through a rpyc request)
requiring a long processing, it delegates it to the main thread, and
returns immediately. When the main thread completes the operation,
it calls bask the hub thread in order to publish the result back to
the client.
The client is modified in order to take this into account (see
walt.client.link). Remote API calls to the server are actually made
into 2 steps:
* send the rpyc request (the hub thread gets it, stores it in a queue
  for the main thread, notifies the main thread, and returns).
* wait for the result by waiting on a response queue.
As a result, there is an important robustness improvement: instead
of waiting for the rpyc call to complete, the client waits for the
result on the queue. Thus, in case an interrupt occurs on the client
side (e.g. ctrl-c), the rpyc call is not broken.
Note: Server-side exceptions occuring when trying to push back the
result although the client is actually disconnected are ignored.

Remaining work: remove other queue mechanisms implemented on the
client side for long operations.